refactor(adapter-dify): extract file handling utilities and clean up debug logs
- Extract file resolution, MIME type mapping, and multimodal serialization logic into separate utils.ts module
- Remove undici dependency from package.json (moved to shared dependencies)
- Clean up all debug/trace logging statements across core and adapter-dify
- Add platform variable to system prompt variables in core/utils/string.ts
- Pass all custom variables from chain to Dify with chatluna_ prefix
- Simplify conversation ID handling by removing redundant checks and temporary workarounds
- Improve code organization and maintainability by separating concerns
