+++ packages/adapter-dify/src/index.ts
@@ -1,9 +1,13 @@
 import { ChatLunaPlugin } from 'koishi-plugin-chatluna/services/chat'
-import { Context, Schema } from 'koishi'
+import { Context, Logger, Schema } from 'koishi'
 import { DifyClientConfig } from './types'
 import { DifyClient } from './client'
+import { createLogger } from 'koishi-plugin-chatluna/utils/logger'
 
+export let logger: Logger
 export function apply(ctx: Context, config: Config) {
+    logger = createLogger(ctx, 'chatluna-dify-adapter')
+
     ctx.on('ready', async () => {
         const plugin = new ChatLunaPlugin<DifyClientConfig, Config>(
             ctx,
