+++ packages/adapter-dify/src/types.ts
@@ -22,3 +22,16 @@ export interface InputFileObject {
     transfer_method: 'local_file'
     upload_file_id: string
 }
+
+export type UploadCandidate = {
+    source: string | ArrayBuffer | Uint8Array | Buffer
+    type: InputFileObject['type']
+    fileName?: string
+    mimeType?: string
+}
+
+export type FilePayload = {
+    buffer: Buffer
+    fileName: string
+    mimeType?: string
+}
